Resolve symlinks
Atom resolves symlinks when opening a file so when working within a symlinked directory path.relative doesn't work and paths in the tree are not the same as in buffers, causing uploadOnSave and right-clicking within the buffer to fail. Might as well just resolve the project path and all file paths.