Add .ssh directory to default ignore paths
Many people use remote-sync for a home directory and this change avoids having .ssh directories downloaded since it contains sensitive SSH keys. I understand if this PR gets denied since not everyone versions a home directory, but I feel it's a small amount of additional work that can help a lot of people.